Move received threads out of MessageContext
